Fixed temporarity during metric conversion, adapted changes to the new metric structure
