+++ test/e2e/testkit/otlp/matchers/metric_matchers.go
@@ -1,3 +1,5 @@
+//go:build e2e
+
 package matchers
 
 import (
