+++ test/e2e/testkit/otlp/metrics/metrics.go
@@ -1,3 +1,5 @@
+//go:build e2e
+
 package metrics
 
 import (
